Rename contextutil to workflowcontext
contextutil was too ambiguous for how closely related to workflow
execution it is. While workflowcontext creates an undesirable directory
hierarchy, it provides clearity for consumers when leveraging the
behavior it provides.
