+++ pkg/workflow/workflowcontext/contextutil.go
@@ -1,5 +1,5 @@
 /*
-Package contextutil contains utility functions for populating workflow context specific data
+Package workflowcontext contains utility functions for populating workflow context specific data
 in a context.Context.
 
 Data appropriate for the context includes anything that cannot be determined at time of
@@ -7,4 +7,4 @@ object construction. For example, a bootstrap cluster does not exist when execut
 workflows, therefore a Kubeconfig isn't available to communicate with the cluster so must be passed
 as contextual data.
 */
-package contextutil
+package workflowcontext
